move unused tsx to tutorial
