+++ app/protected/page.tsx
@@ -1,8 +1,8 @@
-import DeployButton from "@/components/DeployButton";
-import AuthButton from "@/components/AuthButton";
+import DeployButton from "@/components/tutorial/DeployButton";
+import AuthButton from "@/components/tutorial/AuthButton";
 import { createClient } from "@/utils/supabase/server";
 import FetchDataSteps from "@/components/tutorial/FetchDataSteps";
-import Header from "@/components/Header";
+import Header from "@/components/tutorial/Header";
 import { redirect } from "next/navigation";
 
 export default async function ProtectedPage() {
